refactor: extract limiter derivation to separate module
  - Move limiter ceiling constants, deriveLimiterAndPreGain(), buildPreLimiterPrefix(), buildBrickwallLimiter(), limiterPlan,
  LimiterDiagnostics, and planLimiterForLoudnorm() from normalise.go to new limiter.go; no behaviour change
  - Rename sanitizeConfig/sanitizeFloat and related functions to sanitiseConfig/sanitiseFloat across adaptive.go, adaptive_math.go,
  adaptive_test.go, filters.go, filters_test.go (British English spelling)
  - Add fmtLU metric format in metricrow.go for LU loudness range display with 2-decimal places; regenerate golden report with LU
  values displayed as 2 decimals
  - Remove dead default case from formatByRule; hoist analysis TUI styles off View() in analysis_model.go (analysisStatusBox,
  analysisFileStyle, analysisActiveIcon, analysisDoneStyle, analysisErrorStyle, analysisStarStyle, analysisLabelStyle); add safety
  comments for aliased array in-place writes
  - Update docs references: limiter functions now in limiter.go, loudnormTP bounds remain in normalise.go
  (Levelator-Comparison-And-Gap-Analysis.md, Normalisation-Tuning.md)
  - Fix test comment (roundtrip_test.go) and render_test.go helper docstring clarity; remove duplicate close() in roundtrip_test.go
  - Delete four spike scripts from scripts/ (afftdn-spike.sh, anlmdn-matrix-spike.sh, anlmdn-spike.sh, mcompand-spike.sh); user
  decision no longer relevant
  - Pin golangci-lint to v2.12.2 in CI (builder.yml); update release notes installation path to ~/.local/bin (matching
  justfile/README)
  - Add Harper dictionary terms (sorted): acompressor, adeclick, agate, alimiter, aresample, arm64, bandlimit, behaviour,
  brickwall, BS.1770, Bubbletea, cancelled, catalogue, CGO, dependabot, devShell, DI, downmix, ebur128, flavours, golangci,
  highpass, justfile, levelling, loudnorm, lowpass, LU, md, neighbours, normaliser, Otsu, podcast, podcasters, preflight,
  reviewable, RMS, sanitise, sanitiser, sibilant, SPDX, uncached, v2
  - Add integration-test carve-out rule to AGENTS.md testing section: //go:build integration tagged files may decode testdata/
  audio and use fixture helpers; just test and CI skip these files
  - Enforce Dependabot Nix entry preservation rule in AGENTS.md commit guidelines
