+++ .github/workflows/builder.yml
@@ -39,6 +39,8 @@ jobs:
           go install github.com/gordonklaus/ineffassign@v0.2.0
           ineffassign ./...
       - uses: golangci/golangci-lint-action@v9
+        with:
+          version: v2.12.2
 
   lint-actions:
     name: Lint Action ⚙️
@@ -259,16 +261,16 @@ jobs:
 
           ## Installation
 
-          Download the appropriate binary for your platform below, make it executable, and move it to your PATH:
+          Download the appropriate binary for your platform below, make it executable, and move it to `~/.local/bin` (on PATH for most shells, no root required):
 
           ```bash
           # Linux (amd64)
           chmod +x jive-vocals-linux-amd64
-          sudo mv jive-vocals-linux-amd64 /usr/local/bin/jive-vocals
+          mkdir -p ~/.local/bin && mv jive-vocals-linux-amd64 ~/.local/bin/jive-vocals
 
           # macOS (Apple Silicon)
           chmod +x jive-vocals-darwin-arm64
-          sudo mv jive-vocals-darwin-arm64 /usr/local/bin/jive-vocals
+          mkdir -p ~/.local/bin && mv jive-vocals-darwin-arm64 ~/.local/bin/jive-vocals
           ```
 
           ## Checksums
