+++ .harper-dictionary.txt
@@ -1,52 +1,73 @@
-jive-vocals
-Levelator
-VAD
-RMS
-LUFS
-dBTP
-dBFS
-LRA
+acompressor
+adeclick
 afftdn
+agate
+alimiter
 anlmdn
-loudnorm
-ebur128
+aresample
+arm64
 aspectralstats
 astats
-alimiter
-adeclick
-aresample
-acompressor
-agate
-Otsu
 Auphonic
-iZotope
-Riverside
-Zencastr
-Kaye
-libsndfile
+bandlimit
+behaviour
+brickwall
+BS.1770
+Bubbletea
+cancelled
+catalogue
+CGO
+dBFS
+dBTP
 de-esser
 de-essing
-SpeechProfile
-NoiseProfile
-GateSeparationDB
-VoicedLowPercentile
-NoiseHighPercentile
-flooredFraction
-momentary-LUFS
-BS.1770
-EBU
-TidBITS
-MacUpdater
-VideoHelp
-arm64
-NaN
-normaliser
-levelling
-highpass
-lowpass
-bandlimit
-brickwall
+dependabot
+devShell
+DI
 downmix
-sibilant
+EBU
+ebur128
+flavours
+flooredFraction
+GateSeparationDB
+golangci
+highpass
+iZotope
+jive-vocals
+justfile
+Kaye
+Levelator
+levelling
+libsndfile
+loudnorm
+lowpass
+LRA
+LU
+LUFS
+MacUpdater
+md
+momentary-LUFS
+NaN
+neighbours
+NoiseHighPercentile
+NoiseProfile
+normaliser
+Otsu
 podcast
 podcasters
+preflight
+reviewable
+Riverside
+RMS
+sanitise
+sanitiser
+sibilant
+SPDX
+SpeechProfile
+TidBITS
+uncached
+v2
+VAD
+VideoHelp
+VoicedLowPercentile
+Zencastr
