+++ internal/processor/adaptive_math.go
@@ -2,8 +2,8 @@ package processor
 
 import "math"
 
-// sanitizeFloat returns defaultVal if val is NaN or Inf
-func sanitizeFloat(val, defaultVal float64) float64 {
+// sanitiseFloat returns defaultVal if val is NaN or Inf
+func sanitiseFloat(val, defaultVal float64) float64 {
 	if math.IsNaN(val) || math.IsInf(val, 0) {
 		return defaultVal
 	}
